wire up all available nodes
