+++ packages/paperclip-language-service/src/language-service.ts
@@ -11,24 +11,10 @@ Considerations:
 
 import { EngineDelegate, getEngineImports } from "@paperclip-ui/core";
 import { Observable } from "@paperclip-ui/common";
-import {
-  DependencyNodeContent,
-  getAttributeStringValue,
-  getParts as getComponents,
-  hasAttribute,
-  isPaperclipFile,
-  LoadedPCData,
-} from "@paperclip-ui/utils";
 import { AutocompleteService } from "./autocomplete";
 import { collectASTInfo, ColorInfo } from "./collect-ast-info";
 import { DiagnosticService, LintInfo } from "./error-service";
-import { ALL_TAG_NAMES } from "./tag-name-constants";
-import {
-  AvailableElement,
-  AvailableInstance,
-  AvailableNode,
-  AvailableNodeKind,
-} from "./state";
+import { getAllAvailableNodes } from "./state";
 
 export class PaperclipLanguageService {
   private _autocomplete: AutocompleteService;
@@ -80,8 +66,8 @@ export class PaperclipLanguageService {
    * returns all available nodes in the project (text, native elements, custom components)
    */
 
-  getAllAvailableNodes(options: GetAllAvailableNodesOptions) {
-    return getAllAvailableNodes(options, this._engine);
+  getAllAvailableNodes() {
+    return getAllAvailableNodes(this._engine);
   }
 
   /**
@@ -108,56 +94,3 @@ export class PaperclipLanguageService {
     );
   }
 }
-
-const AVILABLE_NATIVE_NODES: AvailableNode[] = [
-  {
-    kind: AvailableNodeKind.Text,
-    name: "text",
-    displayName: "Text",
-    description: "Native text",
-  },
-  ...ALL_TAG_NAMES.map((tagName) => {
-    return {
-      kind: AvailableNodeKind.Element,
-      displayName: tagName,
-      name: tagName,
-      description: "",
-    } as AvailableElement;
-  }),
-];
-
-type GetAllAvailableNodesOptions = {
-  activeDocumentUri?: string;
-};
-
-const getAllAvailableNodes = (
-  { activeDocumentUri }: GetAllAvailableNodesOptions,
-  engine: EngineDelegate
-) => {
-  const allData = engine.getAllLoadedData();
-  const instances: AvailableInstance[] = [];
-
-  for (const uri in allData) {
-    const ast = engine.getLoadedAst(uri) as DependencyNodeContent;
-    if (!isPaperclipFile(uri)) {
-      continue;
-    }
-    const exportedComponents = getComponents(ast).filter((component) =>
-      hasAttribute("export", component)
-    );
-    instances.push(
-      ...exportedComponents.map((component) => {
-        const displayName = getAttributeStringValue("as", component) as string;
-        return {
-          kind: AvailableNodeKind.Instance,
-          displayName,
-          name: displayName,
-          description: "",
-          sourceUri: uri,
-        } as AvailableInstance;
-      })
-    );
-  }
-
-  return [...instances, ...AVILABLE_NATIVE_NODES];
-};
